+++ Src/main.c
@@ -108,8 +108,8 @@ int main(void)
   };
   uint8_t buff[20] = {};
   MPU6050_Status success = MPU6050_init(&hi2c2);
-  uint8_t gyro[6] = {};
-  uint8_t accel[6] = {};
+  float gyro[6] = {};
+  float accel[6] = {};
 
   //0x75 = whoami
 
@@ -121,7 +121,7 @@ int main(void)
   {
     /* USER CODE END WHILE */
     MPU6050_ReadAll(&hi2c2, accel, gyro);
-    sprintf(buff, "A:%d %d %d G:%d %d %d", accel[0], accel[1], accel[2], gyro[0], gyro[1], gyro[2]);
+    sprintf(buff, "A:%f %f %f G:%f %f %f", accel[0], accel[1], accel[2], gyro[0], gyro[1], gyro[2]);
     HAL_UART_Transmit(&huart3, buff, strlen(buff), HAL_MAX_DELAY);
     HAL_Delay(1000);
